+++ packages/hydrogen/src/framework/config.ts
@@ -1,0 +1,15 @@
+declare module globalThis {
+  var __config: HydrogenConfig | undefined;
+}
+
+type HydrogenConfig = {
+  dev?: boolean;
+};
+
+export function setConfig(config: HydrogenConfig) {
+  globalThis.__config = config;
+}
+
+export function getConfig(): HydrogenConfig {
+  return globalThis.__config || {};
+}